+++ R/tt_toString.R
@@ -277,7 +277,8 @@ setMethod("matrix_form", "VTableTree",
                     subtitles = subtitles(obj),
                     page_titles = page_titles(obj),
                     main_footer = main_footer(obj),
-                    prov_footer = prov_footer(obj)
+                    prov_footer = prov_footer(obj),
+                    table_inset = table_inset(obj)
                     )
 })